+++ askbot/doc/source/changelog.rst
@@ -1,8 +1,11 @@
 Changes in Askbot
 =================
 
-Development version
--------------------
+0.10.1 (Nov 16, 2016)
+---------------------
+* Added settings `ASKBOT_ALLOWED_HTML_ELEMENTS` and
+  `ASKBOT_ALLOWED_HTML_ATTRIBUTES`
+* Allow custom markdown parser via `ASKBOT_MARKDOWN_CLASS` setting
 * Applied Akismet check (if enabled) to editing of all content
 * Updated Facebook API to v2.2.
 * Unsubscribe link feature